refactor: Isolate HTTP server concern from `moved` crate into new `server` crate
