+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "moved/src/tests/optimism"]
-	path = moved/src/tests/optimism
+[submodule "server/src/tests/optimism"]
+	path = server/src/tests/optimism
 	url = https://github.com/ethereum-optimism/optimism
