+++ Cargo.toml
@@ -4,11 +4,13 @@ resolver = "2"
 members = [
     "genesis-builder",
     "moved",
+    "server",
 ]
 
 default-members = [
     "genesis-builder",
     "moved",
+    "server",
 ]
 
 [workspace.package]
@@ -50,6 +52,7 @@ move-table-extension = { git = "https://github.com/aptos-labs/aptos-core", tag =
 move-vm-runtime = { git = "https://github.com/aptos-labs/aptos-core", tag = "aptos-node-v1.14.0" }
 move-vm-test-utils = { git = "https://github.com/aptos-labs/aptos-core", tag = "aptos-node-v1.14.0", features = ["table-extension"] }
 move-vm-types = { git = "https://github.com/aptos-labs/aptos-core", tag = "aptos-node-v1.14.0" }
+moved = { path = "moved" }
 once_cell = "1.19"
 openssl = "0.10"
 serde = { version = "1", features = ["derive"] }
