+++ moved/Cargo.toml
@@ -10,7 +10,6 @@ alloy-eips.workspace = true
 alloy-primitives.workspace = true
 alloy-rlp.workspace = true
 alloy-serde.workspace = true
-anyhow.workspace = true
 aptos-crypto.workspace = true
 aptos-framework.workspace = true
 aptos-gas-meter.workspace = true
@@ -23,11 +22,7 @@ aptos-jellyfish-merkle.workspace = true
 aptos-storage-interface.workspace = true
 bcs.workspace = true
 bytes.workspace = true
-clap.workspace = true
-flate2.workspace = true
 hex.workspace = true
-hyper.workspace = true
-jsonwebtoken.workspace = true
 move-binary-format.workspace = true
 move-core-types.workspace = true
 move-table-extension.workspace = true
@@ -42,12 +37,9 @@ sui-framework.workspace = true
 sui-types.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
-warp-reverse-proxy.workspace = true
-warp.workspace = true
 
 [dev-dependencies]
 alloy.workspace = true
-dotenvy.workspace = true
-openssl.workspace = true
+anyhow.workspace = true
 move-compiler.workspace = true
 test-case.workspace = true
