+++ server/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "moved-server"
+description = "Moved execution HTTP server (using warp)"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+aptos-types.workspace = true
+bcs.workspace = true
+clap.workspace = true
+flate2.workspace = true
+hex.workspace = true
+hyper.workspace = true
+jsonwebtoken.workspace = true
+move-core-types.workspace = true
+moved.workspace = true
+once_cell.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+warp-reverse-proxy.workspace = true
+warp.workspace = true
+
+[dev-dependencies]
+alloy.workspace = true
+dotenvy.workspace = true
+openssl.workspace = true
