+++ server/src/tests/config.sh
