+++ server/src/tests/res/ERC20.json
